backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
redis==5.0.4
websockets==12.0
openai==1.28.0
python-dotenv==1.0.1
aiohttp==3.9.5
httpx==0.27.2